Move Analytics class to common namespace
