+++ src/common/analytics.h
