+++ src/common/common.pro
@@ -18,7 +18,9 @@ QMAKE_CXXFLAGS += -std=c++11
 INCLUDEPATH += $$top_srcdir/src
 
 SOURCES += downloadmanager.cpp \
-           logger.cpp
+           logger.cpp          \
+           analytics.cpp
 
 HEADERS += downloadmanager.h \
-           logger.h
+           logger.h          \
+           analytics.h
