+++ src/gui/mzroll/alignmentdialog.cpp
@@ -1,7 +1,7 @@
 #include <QDesktopWidget>
 
 #include "alignmentdialog.h"
-#include "analytics.h"
+#include "common/analytics.h"
 #include "background_peaks_update.h"
 #include "Compound.h"
 #include "globals.h"
